# Allow better caching of SPA launch pages