--- a/kolibri/core/assets/src/utils/userAgent.js
+++ b/kolibri/core/assets/src/utils/userAgent.js
@@ -1,3 +1,10 @@
+/*
+ * userAgent regex is imported from here:
+ * https://github.com/faisalman/ua-parser-js/blob/master/src/ua-parser.js
+ * The subsequent code is then written from scratch as it was easier than vendoring
+ * the more general purpose code therein.
+ */
+
 const NAME = 'name';
 const VERSION = 'version';
 
@@ -49,10 +56,37 @@ const browserTests = [
   },
 ];
 
+/**
+ * A browser specification object.
+ * All version numbers are actually strings to allow for non-final version numbers.
+ * @typedef {Object} Browser.
+ * @property {?string} name - A nullable string that gives the name of the recognized browser.
+ * @property {?string} major - Major version number.
+ * @property {?string} minor - Minor version number.
+ * @property {?string} patch - Patch version number.
+ */
+
+/**
+ * This function parses the browser user agent string and returns an object with two values
+ * one for the recognized browser name, and one for the recognized browser version.
+ * It relies on the regexes above to infer the browser and version from user agent string.
+ * As such, its ability to recognizes browsers is limited, as it will only return non-null
+ * values for browser and version in the case that the user agent matches those above.
+ * @param  {string} userAgent a user agent string from a browser's window.navigator.userAgent.
+ * @return {Browser} - a browser specification object.
+ */
 export function getBrowser(userAgent) {
+  // Use to track the parsed name and version.
   const browser = {
     [NAME]: null,
     [VERSION]: null,
+  };
+  // Setup the output object which will be returned.
+  const outputBrowser = {
+    name: null,
+    major: null,
+    minor: null,
+    patch: null,
   };
 
   if (userAgent) {
@@ -88,24 +122,51 @@ export function getBrowser(userAgent) {
       }
     }
   }
-  if (browser[VERSION]) {
-    browser[VERSION] = browser[VERSION].split('.');
+  if (browser[NAME]) {
+    outputBrowser.name = browser[NAME];
   }
-  return browser;
+
+  if (browser[VERSION]) {
+    const version = browser[VERSION].split('.');
+    if (version[0]) {
+      outputBrowser.major = version[0];
+      if (version[1]) {
+        outputBrowser.minor = version[1];
+        if (version[2]) {
+          outputBrowser.patch = version[2];
+        }
+      }
+    }
+  }
+  return outputBrowser;
 }
 
+/**
+ * A requirements specification object.
+ * @typedef {Object.<string, Browser>} Requirements.
+ */
+
+/**
+ * A function to determine if the browser specification object fails to pass one of the restrictions
+ * passed in. It does this in a permissive way, whereby if no specification is made for a particular
+ * browser name, then it will return true. If, however, the browser is specified in the requirements
+ * object, then if it fails to pass the version requirements specified, then false will be returned.
+ * @param  {Browser} browser     - a browser specification object for the browser being tested.
+ * @param  {Requirements} requirements - a requirements specification object specifying conditions
+ * to test.
+ * @return {boolean} - false if failed to pass any of the requirements, true otherwise.
+ */
 export function passesRequirements(browser, requirements) {
-  if (browser[VERSION] && browser[NAME]) {
-    const entry = requirements[browser[NAME]];
+  if (browser.major && browser.name) {
+    const entry = requirements[browser.name];
     if (entry) {
-      const version = browser[VERSION];
-      if (version[0] < entry.major) {
+      if (browser.major < entry.major) {
         return false;
-      } else if (version[0] === entry.major) {
-        if (entry.minor && (version[1] < entry.minor || !version[1])) {
+      } else if (browser.major === entry.major) {
+        if (entry.minor && (browser.minor < entry.minor || !browser.minor)) {
           return false;
-        } else if (entry.minor && version[1] === entry.minor) {
-          if (entry.patch && (version[2] < entry.patch || !version[2])) {
+        } else if (entry.minor && browser.minor === entry.minor) {
+          if (entry.patch && (browser.patch < entry.patch || !browser.patch)) {
             return false;
           }
         }