--- a/kolibri/core/context_processors/custom_context_processor.py
+++ b/kolibri/core/context_processors/custom_context_processor.py
@@ -25,13 +25,12 @@ def pass_browser_entry(agent, entry):
 
 
 def supported_browser(request):
-    if "supported_browser" not in request.session:
-
-        user_agent = user_agents.parse(request.META.get("HTTP_USER_AGENT", ""))
-        request.session["supported_browser"] = all(
+    user_agent = user_agents.parse(request.META.get('HTTP_USER_AGENT', ''))
+    return {
+        'supported_browser': all(
             pass_browser_entry(user_agent, entry) for entry in browser_requirements
         )
-    return {"supported_browser": request.session["supported_browser"]}
+    }
 
 
 def developer_mode(request):