--- a/kolibri/core/device/middleware.py
+++ b/kolibri/core/device/middleware.py
@@ -1,15 +1,58 @@
+from django.conf import settings
 from django.http import HttpResponse
-from django.middleware.locale import LocaleMiddleware
+from django.http import HttpResponseRedirect
+from django.urls import is_valid_path
 from django.utils import translation
 
 from .translation import get_language_from_request
+from kolibri.utils.conf import OPTIONS
 
 
-class KolibriLocaleMiddleware(LocaleMiddleware):
-    def process_request(self, request):
-        language = get_language_from_request(request)
+class KolibriLocaleMiddleware:
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        language, language_from_path = get_language_from_request(request)
         translation.activate(language)
         request.LANGUAGE_CODE = translation.get_language()
+
+        response = self.get_response(request)
+
+        language = translation.get_language()
+
+        urlconf = getattr(request, "urlconf", settings.ROOT_URLCONF)
+
+        if response.status_code == 404 and not language_from_path:
+            # Maybe the language code is missing in the URL? Try adding the
+            # language prefix and redirecting to that URL.
+            script_prefix = OPTIONS["Deployment"]["URL_PATH_PREFIX"].strip("/")
+            if script_prefix:
+                language_path = request.path_info.replace(script_prefix, "%s/%s" % (script_prefix, language))
+            else:
+                language_path = "/" + language + request.path_info
+            path_valid = is_valid_path(language_path, urlconf)
+            path_needs_slash = (
+                not path_valid and (
+                    settings.APPEND_SLASH and not language_path.endswith("/")
+                    and is_valid_path("%s/" % language_path, urlconf)
+                )
+            )
+            if path_valid or path_needs_slash:
+                script_prefix = OPTIONS["Deployment"]["URL_PATH_PREFIX"]
+                # Insert language after the script prefix and before the
+                # rest of the URL
+                language_url = request.get_full_path(force_append_slash=path_needs_slash).replace(
+                    script_prefix,
+                    "%s%s/" % (script_prefix, language),
+                    1
+                )
+                return HttpResponseRedirect(language_url)
+
+        if "Content-Language" not in response:
+            response["Content-Language"] = language
+
+        return response
 
 
 class IgnoreGUIMiddleware(object):