--- a/kolibri/core/device/middleware.py
+++ b/kolibri/core/device/middleware.py
@@ -1,5 +1,4 @@
 from django.conf import settings
-from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from django.urls import is_valid_path
 from django.utils import translation
@@ -62,25 +61,3 @@ class KolibriLocaleMiddleware(object):
                 response["Content-Language"] = language
 
         return response
-
-
-class IgnoreGUIMiddleware(object):
-    def __init__(self, get_response):
-        self.get_response = get_response
-        # One-time configuration and initialization.
-
-    def __call__(self, request):
-        # Code to be executed for each request before
-        # the view (and later middleware) are called.
-
-        response = self.get_response(request)
-
-        # Code to be executed for each request/response after
-        # the view is called.
-
-        return response
-
-    def process_view(self, request, view_func, view_args, view_kwargs):
-        if request.META.get("HTTP_USER_AGENT", None) == "Kolibri session":
-            return HttpResponse("")
-        return None