… kolibri/core/assets/src/utils/__test__/userAgent.spec.js
… kolibri/core/device/test/locale_middleware_urls.py
… kolibri/core/device/test/prefixed_locale_middleware_urls.py
… kolibri/core/device/test/test_locale_middleware.py
… kolibri/core/kolibri_plugin.py
… kolibri/core/templates/kolibri/base.html
… kolibri/core/templates/kolibri/unsupported_browser.html
… kolibri/core/templatetags/kolibri_tags.py
… kolibri/core/test/test_context_processor.py
… kolibri/core/test/test_key_urls.py
… kolibri/core/test/test_setlanguage.py
… kolibri/core/urls.py
… kolibri/core/views.py
… kolibri/deployment/default/settings/dev.py
… kolibri/plugins/base.py
… kolibri/plugins/coach/kolibri_plugin.py
… kolibri/plugins/coach/urls.py
… kolibri/plugins/coach/views.py
… kolibri/plugins/demo_server/urls.py
… kolibri/plugins/device_management/kolibri_plugin.py
… kolibri/plugins/device_management/urls.py
… kolibri/plugins/device_management/views.py
… kolibri/plugins/facility_management/kolibri_plugin.py
… kolibri/plugins/facility_management/urls.py
… kolibri/plugins/facility_management/views.py
… kolibri/plugins/learn/assets/src/apiResources.js
… kolibri/plugins/learn/assets/src/views/LearnSideNavEntry.vue
… kolibri/plugins/learn/kolibri_plugin.py
… kolibri/plugins/learn/templates/learn/learn.html
… kolibri/plugins/learn/test/test_learner_classroom.py