@rtibbles 2019-02-12
We can hold off on this until 0.12.1 - I mostly want to have a few things in our pocket to allow performance improvements if we discover that we need them for 0.12.0.