@indirectlylit 2019-02-13
setting milestone to 0.12.1

note there's a conflict in kolibri_app.js