@rtibbles 2019-02-13
Not a hugely representative methodology, but I ran siege with default settings (25 concurrent users), and these are the results on current release-v0.12.x:
```
Transactions:		        1000 hits
Availability:		      100.00 %
Elapsed time:		       65.42 secs
Data transferred:	        6.90 MB
Response time:		        1.60 secs
Transaction rate:	       15.29 trans/sec
Throughput:		        0.11 MB/sec
Concurrency:		       24.53
Successful transactions:        1000
Failed transactions:	           0
Longest transaction:	        5.91
Shortest transaction:	        0.18
```

These are the results when run on this PR:
```
Transactions:		        1000 hits
Availability:		      100.00 %
Elapsed time:		       10.00 secs
Data transferred:	        6.76 MB
Response time:		        0.25 secs
Transaction rate:	      100.00 trans/sec
Throughput:		        0.68 MB/sec
Concurrency:		       24.72
Successful transactions:        1000
Failed transactions:	           0
Longest transaction:	        3.69
Shortest transaction:	        0.06
```

Testing conditions definitely favour this PR and what it is doing, but it is still a roughly 6 times speed up on average response time.