@rtibbles 2019-02-13
The other thing to note, obviously, is that this is moving some bootstrapping out of user pages, so this will result in more requests to the server, so we would probably want to do a live load test to properly assess the impact of this.