@rtibbles 2019-03-28
Note that a live load test of this showed improved performance of this over current release 0.12 - although there was a cache warm-up phase where things were a little slower.