@rtibbles 2019-03-29
This PR now also removes the GZip middleware which seems to have been a considerable performance bottleneck on our bulkier API endpoints. As an example, here is a siege run against the channel endpoint, with and without GZip:
```
Channel Endpoint:
With GZip:
Transactions:		         500 hits
Availability:		      100.00 %
Elapsed time:		       10.04 secs
Data transferred:	       80.57 MB
Response time:		        0.49 secs
Transaction rate:	       49.80 trans/sec
Throughput:		        8.03 MB/sec
Concurrency:		       24.63
Successful transactions:         500
Failed transactions:	           0
Longest transaction:	        2.27
Shortest transaction:	        0.08

Without GZip:
Transactions:		         500 hits
Availability:		      100.00 %
Elapsed time:		        5.79 secs
Data transferred:	      111.32 MB
Response time:		        0.24 secs
Transaction rate:	       86.36 trans/sec
Throughput:		       19.23 MB/sec
Concurrency:		       20.75
Successful transactions:         500
Failed transactions:	           0
Longest transaction:	        3.02
Shortest transaction:	        0.01
```