@rtibbles 2019-04-02
Thanks - I'll double check the language switching behaviour from the frontend some more, and see if I can replicate.