@rtibbles 2019-04-15
Have now addressed all comments - the language changing behaviour was indeed broken. Have fixed the issue, and also reimplemented redirect to make sure that when you change language it returns you to exactly the same page (including fragment and query param) that you were on previously.

I think this is ready to merge, assuming there are no tests that I have missed.